Trekker ut enhetssøk til egen komponent
